Authen and Author/requirements.txt
Flask
pillow
reportlab
PyJWT
bcrypt
SQLAlchemy